refactor: move result summary into it's own module
