+++ .github/workflows/summarize-results.yml
@@ -28,10 +28,10 @@ jobs:
 
       - name: Summarize evaluation results
         run: |
-          uv run bcbench evaluate summarize --result-dir "${{ inputs.results-dir }}"
+          uv run bcbench result summarize --result-dir "${{ inputs.results-dir }}"
         shell: pwsh
 
-      - name: Upload evaluation summary
+      - name: Upload evaluation summary to artifacts
         uses: actions/upload-artifact@v4
         with:
           name: evaluation-summary
