+++ src/bcbench/results/__init__.py
@@ -1,0 +1,4 @@
+from bcbench.results.display import create_console_summary, create_github_job_summary
+from bcbench.results.evaluation_result import EvaluationResult, EvaluationResultSummary
+
+__all__ = ["EvaluationResult", "EvaluationResultSummary", "create_console_summary", "create_github_job_summary"]
